// vaccineApp/src/infrastructure/interfaces/consult-vaccine-response.ts
// Generated by https://quicktype.io

export interface ConsultByIndependentResponse {
    apply_vaccines: ApplyVaccine[];
    desde:          string;
    limite:         string;
    resp:           boolean;
    statusCode:     number;
    total:          number;
}

export interface ApplyVaccine {
    _id:              ID;
    dependent:        Dependent;
    dosis:            Dosis;
    image:            string;
    lote:             string;
    status:           boolean;
    vaccination_date: string;
}

export interface ID {
    $oid: string;
}

export interface Dependent {
    _id:             ID;
    age:             number;
    birth:           string;
    city:            string;
    days_birth:      number;
    email:           string;
    gender_id:       string;
    isChildren:      boolean;
    isUser:          boolean;
    lastname:        string;
    name:            string;
    phone:           string;
    relationship_id: string;
    state:           string;
    status:          boolean;
    user_id:         string;
}

export interface Dosis {
    _id:             ID;
    age_frequency:   string;
    columReporte:    string;
    expires_in_days: number;
    name:            string;
    rowReporte:      string;
    status:          boolean;
    vaccine:         Vaccine;
}

export interface Vaccine {
    _id:              ID;
    application_age:  string;
    description:      string;
    disease_prevents: string;
    isChildren:       boolean;
    name:             string;
    status:           boolean;
}